legacy_caregiver_application/requirements.txt
fhir.resources==6.1.0
Flask==2.0.1
requests==2.23.0